apply base filters when no tenant selected, but user is assigned to tenants
